Last part of removing barrel files
