+++ apps/studio/components/interfaces/ProjectAPIDocs/Content/Introduction.tsx
@@ -11,7 +11,7 @@ import ContentSnippet from '../ContentSnippet'
 import { DOCS_CONTENT } from '../ProjectAPIDocs.constants'
 import type { ContentProps } from './Content.types'
 
-const Introduction = ({ showKeys, language, apikey, endpoint }: ContentProps) => {
+export const Introduction = ({ showKeys, language, apikey, endpoint }: ContentProps) => {
   const { ref } = useParams()
   const { data: apiKeys } = useAPIKeysQuery({ projectRef: ref })
   const { data } = useProjectSettingsV2Query({ projectRef: ref })
@@ -139,5 +139,3 @@ const Introduction = ({ showKeys, language, apikey, endpoint }: ContentProps) =>
     </>
   )
 }
-
-export default Introduction
