+++ apps/studio/components/interfaces/ProjectAPIDocs/Content/Storage.tsx
@@ -2,12 +2,10 @@ import ContentSnippet from '../ContentSnippet'
 import { DOCS_CONTENT } from '../ProjectAPIDocs.constants'
 import type { ContentProps } from './Content.types'
 
-const Storage = ({ language }: ContentProps) => {
+export const Storage = ({ language }: ContentProps) => {
   return (
     <>
       <ContentSnippet selectedLanguage={language} snippet={DOCS_CONTENT.storage} />
     </>
   )
 }
-
-export default Storage
