+++ apps/studio/components/interfaces/ProjectAPIDocs/Content/UserManagement.tsx
@@ -3,7 +3,7 @@ import ContentSnippet from '../ContentSnippet'
 import { DOCS_CONTENT } from '../ProjectAPIDocs.constants'
 import type { ContentProps } from './Content.types'
 
-const UserManagement = ({ language, apikey, endpoint }: ContentProps) => {
+export const UserManagement = ({ language, apikey, endpoint }: ContentProps) => {
   const { authenticationSignInProviders } = useIsFeatureEnabled([
     'authentication:sign_in_providers',
   ])
@@ -87,5 +87,3 @@ const UserManagement = ({ language, apikey, endpoint }: ContentProps) => {
     </>
   )
 }
-
-export default UserManagement
