--- apps/studio/components/interfaces/ProjectAPIDocs/Content/index.ts
@@ -1,12 +1,0 @@
-export { default as EdgeFunctions } from './EdgeFunctions'
-export { default as Entities } from './Entities'
-export { default as Introduction } from './Introduction'
-export { default as Realtime } from './Realtime'
-export { default as Storage } from './Storage'
-export { default as StoredProcedures } from './StoredProcedures'
-export { default as UserManagement } from './UserManagement'
-
-export { default as Bucket } from './Bucket'
-export { default as EdgeFunction } from './EdgeFunction'
-export { default as Entity } from './Entity'
-export { RPC } from './RPC'
