+++ apps/studio/components/interfaces/ProjectAPIDocs/ProjectAPIDocs.tsx
@@ -6,19 +6,17 @@ import { getKeys, useAPIKeysQuery } from 'data/api-keys/api-keys-query'
 import { useProjectSettingsV2Query } from 'data/config/project-settings-v2-query'
 import { useCustomDomainsQuery } from 'data/custom-domains/custom-domains-query'
 import { useAppStateSnapshot } from 'state/app-state'
-import {
-  Bucket,
-  EdgeFunction,
-  EdgeFunctions,
-  Entities,
-  Entity,
-  Introduction,
-  RPC,
-  Realtime,
-  Storage,
-  StoredProcedures,
-  UserManagement,
-} from './Content'
+import { Bucket } from './Content/Bucket'
+import { EdgeFunction } from './Content/EdgeFunction'
+import { EdgeFunctions } from './Content/EdgeFunctions'
+import { Entities } from './Content/Entities'
+import { Entity } from './Content/Entity'
+import { Introduction } from './Content/Introduction'
+import { Realtime } from './Content/Realtime'
+import { RPC } from './Content/RPC'
+import { Storage } from './Content/Storage'
+import { StoredProcedures } from './Content/StoredProcedures'
+import { UserManagement } from './Content/UserManagement'
 import FirstLevelNav from './FirstLevelNav'
 import LanguageSelector from './LanguageSelector'
 import SecondLevelNav from './SecondLevelNav'
@@ -34,7 +32,7 @@ import SecondLevelNav from './SecondLevelNav'
  * - GraphiQL needs a better home, cannot be placed under Database as its "API"
  */
 
-const ProjectAPIDocs = () => {
+export const ProjectAPIDocs = () => {
   const { ref } = useParams()
   const snap = useAppStateSnapshot()
   const isIntroduction =
@@ -154,5 +152,3 @@ const ProjectAPIDocs = () => {
     </SidePanel>
   )
 }
-
-export default ProjectAPIDocs
