+++ apps/studio/components/interfaces/Settings/General/DeleteProjectPanel/DeleteProjectPanel.tsx
@@ -2,7 +2,7 @@ import { FormHeader } from 'components/ui/Forms/FormHeader'
 import { useSelectedOrganizationQuery } from 'hooks/misc/useSelectedOrganization'
 import { useSelectedProjectQuery } from 'hooks/misc/useSelectedProject'
 import { Alert_Shadcn_, AlertDescription_Shadcn_, AlertTitle_Shadcn_, CriticalIcon } from 'ui'
-import DeleteProjectButton from './DeleteProjectButton'
+import { DeleteProjectButton } from './DeleteProjectButton'
 
 export const DeleteProjectPanel = () => {
   const { data: project } = useSelectedProjectQuery()
