--- apps/studio/components/interfaces/Settings/General/index.ts
@@ -1,6 +1,0 @@
-export { default as ComplianceConfig } from './ComplianceConfig/ProjectComplianceMode'
-export { CustomDomainConfig } from './CustomDomainConfig/CustomDomainConfig'
-export { default as DeleteProjectButton } from './DeleteProjectPanel/DeleteProjectButton'
-export { default as General } from './General'
-export { default as TransferProjectButton } from './TransferProjectPanel/TransferProjectButton'
-export { default as TransferProjectPanel } from './TransferProjectPanel/TransferProjectPanel'
