+++ apps/studio/lib/helpers.test.ts
@@ -1,4 +1,6 @@
+import { v4 as _uuidV4 } from 'uuid'
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
+
 import {
   detectBrowser,
   detectOS,
@@ -22,9 +24,22 @@ import {
   timeout,
   tryParseInt,
   tryParseJson,
+  uuidv4,
 } from './helpers'
 
 import { copyToClipboard } from 'ui'
+
+vi.mock('uuid', () => ({
+  v4: vi.fn(() => 'mocked-uuid'),
+}))
+
+describe('uuidv4', () => {
+  it('calls uuid.v4 and returns the result', () => {
+    const result = uuidv4()
+    expect(_uuidV4).toHaveBeenCalled()
+    expect(result).toBe('mocked-uuid')
+  })
+})
 
 describe('tryParseJson', () => {
   it('should return the parsed JSON', () => {
